backend/requirements.txt
asyncpg==0.30.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cssselect==1.3.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
greenlet==3.2.2
idna==3.10
jieba3k==0.35.1
joblib==1.5.0
lxml==5.4.0
newspaper3k==0.2.8
nltk==3.9.1
pillow==11.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.7
sqlalchemy==2.0.40
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typing-extensions==4.13.2
urllib3==2.4.0
